exception from touint32 is input -1, now use toint64 and then cast to uint to get 0xffffffff in armt32 assembler

+++ src/Assembler/Commands/ARMT32/Asm/T2LdrConstCommand.cs
@@ -99,7 +99,7 @@ namespace Yama.Assembler.ARMT32
             assembleFormatBranch.Arguments.Add(valueSkip);
 
             RequestAssembleFormat assembleFormatConst = new RequestAssembleFormat();
-            assembleFormatConst.Command = Convert.ToUInt32(t.Argument1.Token.Value);
+            assembleFormatConst.Command = (uint)Convert.ToInt64(t.Argument1.Token.Value);
             assembleFormatConst.Result = assembleFormat.Result;
 
             if (!format.Assemble(assembleFormat)) return false;
